Remove the offline profile cache.
The whole reason for adding this was because of the legacy skin servers
which would take ages (upwards of 3 minutes) to respond. Since those are
now removed (no longer supported), we don't really need it any more.

Using only Valhalla, the skin server responds close to the same time as
Mojang's server. Additionally, since Java16 the JsonWriter has started to
behave erratically and causes the cache thread to block, so it's better to
just remove this feature anyway.